Create README.md
# CardMatchGame

What the project does -
This project is about card matching game. You have to select 1 card then you have to select another card then code will check if they both cards are the same if they are same
then instead of that card a another white card will come. and so on .. users every moved counted and every match counted and shown on the top of the game board and below is the
restart button on clicking on restart all will shuffle randomly and game restarted for you. 

How to run project -
first clone project in your machine. then hit yarn install or npm install. then after that hit yarn start then app will run in your emulator or in your mobile (if your mobile
is connected with your pc with debugging on).